Add files via upload
Added "main" method to begin processing commands as entered by user in the starting room of the labyrinth game. Added "clean" method to add uniformity to all commands entered by user -- all commands are made lowercase, extra spaces are removed, and valid commands are abbreviated to their first letter. Current valid commands are "north", "south", "east", and "examine". Added "examine" method to begin list of targets that can be used with that method. Noted the necessity of a location tracker but have not yet implemented one. Need to add the paper list to "examine" method.