Location class
Added a location class in order to later create an object of said class that can be used to track where a user is in the maze. Also added the "slip of paper" into the "examine" command method. Print statement for it needs to be finished. Also considered that "slip of paper" could easily become a repeat command -- later in development may need to add an additional question to determine which "slip of paper" user is referring to. Might want to create a separate method for directional print statements too?