Starting List
Added the items in the list found in the starting room as a string within the examine function. Realized that the location name will be impossible to access without an instance of the object which cannot be shared across main and examine functions -- troubleshoot this. Added a variable that notates the end of the "while" loop that keeps prompting users for inputs.